fix: only render prototype creator if required components are provided
